added support for inherited styles; bug fixes
- Author Styles now support inherited styles. A mocked parent element is
created with inheritable properties computed.
- Fixed problem where injected stylesheets could be processed out of
order. External stylesheets that are inaccessible due to CORs security
restrictions are injected asynchronously. There's no gaurantee of the
stylesheets being injected in the same order they were processed. So,
I've added attributes to the LINK and STYLE tags to track the order in
which they should be processed.
- Code reuse has been increased and some repetitious code has been
removed.
- Some code has been refactored to use plain JavaScript. A goal is to
remove jQuery dependencies to reduce complexity and improve performance.