bug fix: prevent jquery from conflicting
jQuery is injected into the page for the sake of the sidebar developer
tools page, but it was conflicting with libraries on some pages and
causing pages to malfunction. This update should resolve the conflict by
injecting jQuery only when it does not already exist and it injects it
at the top of the HEAD so that other libraries can override it.
Sidebar.js references jQuery instead of $ in case of multiple libraries.