improved performance
Now using jQuery's is() instead of indexOf() to test an element against
a selector. indexOf() forces jQuery itself to iterate over the
collection returned by the selector to find the element's index, but I
suspect that is() is using the new browser method, matchesSelector().
Either it is simply faster at iterating due to native code or they have
a better way to do the lookup.

I also started adding code for a future method of inspecting an element
that I'll call "Computed Author Styles" which will result in getting the
fewest author styles possible for each element.

I also changed the way inherited styles are calculated. I'm calculating
them myself now by walking through the parent elements in an effort to
return the fewest style declarations possible. The alternative is to
rely on the browser's getComputedStyle() which returns too much
information. We're after the fewest CSS rules that will achieve the same
style. This mimics what we do when authoring styles.